Db is ready.
Now it supports create, rename, unlink, but log is a little bit buggy.
Conflicts should be solved by cache_solve_conflict().

If a file in cache is removed and then created again, remove action
should be changed to create.

If a file is created and then removed, whole entry can be deleted from
log like nothing happend.

If a file is created and contains some data, there is no need to change
create to write, because file will be synced by default.

Another action will be replaced by remove.